Add a drift_tolerance to crawler mode.
Entries should not ever be deleted from the public
database hosted by the WVSOS. But if they are ... it
breaks our crawler mode as we don't know if we're
at the end of the line or just on a deleted entry.

This adds a lookahead / drift tolerance of 3, so
that we check at least 3 entries ahead of a missing
one before ending a crawl session.

This fixes #2.